pyproject.toml
[project]
name = "halligame"
version = "0.0.0alpha"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyrlang",
    "pyrlang-term",
    "psutil",
    "pyfiglet",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/halligame", "src/cli"]

[tool.uv.sources]
pyrlang-term = { path = "vendor/Term" }
pyrlang = { path = "vendor/Pyrlang" }

[tool.ruff]
line-length = 80
indent-width = 4
exclude = [
    ".git",
    ".mypy_cache",
    ".nox",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "venv",
    "vendor",
]

[tool.ruff.lint]
extend-select = [
    "E501", # line too long
    "UP",  # pyupgrade
    "I",   # Import sorting
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"


